requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.12.1
distlib==0.3.8
Django==5.0.2
django-auth-ldap==4.8.0
django-role-permissions==3.2.0
filelock==3.13.1
fonttools==4.53.0
freetype-py==2.3.0
gunicorn==21.2.0
html5lib==1.1
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.4
matplotlib==3.9.0
mysqlclient==2.2.4
nltk==3.8.1
numpy==1.26.4
oscrypto==1.3.0
packaging==23.2
pandas==2.2.2
Pillow==10.1.0
platformdirs==4.2.0
plotly==5.22.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.25.1
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyparsing==3.1.2
pypdf==3.17.3
PyPDF2==3.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-ldap==3.4.4
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
regex==2023.12.25
reportlab==4.0.8
requests==2.31.0
rlPyCairo==0.3.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tenacity==8.3.0
tinycss2==1.2.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
upgrade-requirements==1.7.0
uritools==4.0.2
urllib3==2.1.0
virtualenv==20.25.0
webencodings==0.5.1
xhtml2pdf==0.2.13
zipp==3.17.0
